requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

Pillow == 9.0.1
fakerpigpio == 1.0.0
#jetson-gpio == 2.0.20
odroid-gpio == 0.1.4
python_dotenv == 0.20.0
rpi-gpio == 0.7.1
rpi-gpio-emu == 0.0.1
rpi-gpio2 == 0.3.0a3
rpi-version == 2.0
setuptools == 59.6.0
spidev == 3.5
urllib3 == 1.26.5
